Use ".auth.scheme" for java package
Also, removed/updated some TODOs.
